refactor: change struct field types.
previously, we used `void` using unions. however, this
is inaccurate in a sense that `void` is predominantly used
to indicate that a function returns nothing. `null`
is similar to `void` but it's used to indicate that
a parameter's value is "nothing" or an absence of value.

hence, we are now passing "nullable" fields as `nullable *str`.

Signed-off-by: Soc Virnyl Estela <contact@uncomfyhalomacro.pl>